ml-api/requirements.txt
Flask
gunicorn
numpy
pillow
tensorflow
tensorflow-hub